+++ src/meta/api/src/share_api_test_suite.rs
@@ -25,13 +25,13 @@ use common_meta_types::MetaError;
 use enumflags2::BitFlags;
 use tracing::info;
 
-use crate::get_kv_data;
 use crate::get_object_shared_by_share_ids;
 use crate::get_share_account_meta_or_err;
 use crate::get_share_id_to_name_or_err;
 use crate::get_share_meta_by_id_or_err;
 use crate::get_share_or_err;
 use crate::is_all_db_data_removed;
+use crate::testing::get_kv_data;
 use crate::ApiBuilder;
 use crate::AsKVApi;
 use crate::KVApi;
