+++ src/meta/api/src/util.rs
@@ -20,7 +20,6 @@ use common_meta_app::schema::DatabaseMeta;
 use common_meta_app::schema::DatabaseNameIdent;
 use common_meta_app::schema::TableNameIdent;
 use common_meta_app::share::*;
-use common_meta_types::anyerror::AnyError;
 use common_meta_types::app_error::AppError;
 use common_meta_types::app_error::ShareHasNoGrantedDatabase;
 use common_meta_types::app_error::UnknownDatabase;
@@ -523,28 +522,6 @@ where
         }
     }
     Ok((false, None))
-}
-
-/// Get existing value by key. Panic if key is absent.
-/// This function is only used for testing.
-pub async fn get_kv_data<T>(
-    kv_api: &(impl KVApi + ?Sized),
-    key: &impl KVApiKey,
-) -> Result<T, MetaError>
-where
-    T: FromToProto,
-    T::PB: common_protos::prost::Message + Default,
-{
-    let res = kv_api.get_kv(&key.to_key()).await?;
-    if let Some(res) = res {
-        let s = deserialize_struct(&res.data)?;
-        return Ok(s);
-    };
-
-    Err(MetaError::Fatal(AnyError::error(format!(
-        "failed to get {}",
-        key.to_key()
-    ))))
 }
 
 pub async fn get_object_shared_by_share_ids(
